chore: split out frequency components
