+++ src/components/CheckEditor/FormComponents/Frequency.types.ts
@@ -1,0 +1,6 @@
+export interface FrequencyComponentProps {
+  value: number;
+  onChange: (value: number) => void;
+  min: number;
+  max: number;
+}
