+++ src/components/CheckEditor/FormComponents/Frequency.utils.ts
@@ -1,0 +1,6 @@
+export function frequencyInSecondsAndMinutes(milliseconds: number) {
+  const minutes = Math.floor(milliseconds / 60000);
+  const seconds = Math.floor((milliseconds % 60000) / 1000);
+
+  return { minutes, seconds };
+}
